updating to master tip

+++ powderday/nebular_emission/cloudy_model.py
@@ -1,6 +1,6 @@
 from powderday.nebular_emission.abund import getNebAbunds
 from powderday.nebular_emission.ASCIItools import *
-from powderday.nebular_emission.cloudy_tools import air_to_vac, calc_LogU, convert_metals
+from powderday.nebular_emission.cloudy_tools import air_to_vac, calc_LogQ, convert_metals
 import powderday.config as cfg
 from astropy import constants
 import logging
@@ -33,7 +33,7 @@ def write_input_sed(wav, spec):
     ascii_file = str(uuid.uuid4().hex) + ".ascii"
     while compiled_exists(ascii_file):
         ascii_file = str(uuid.uuid4().hex) + ".ascii"
-
+      
     logging.info("Executing write ascii sequence...")
     logging.info("Writing.....")
     WriteASCII(ascii_file, wav, spec, nx=len(wav), nmod=1, par1_val=1.e6)
@@ -66,11 +66,10 @@ def write_cloudy_input(**kwargs):
             "dens": 100.0,     # number density of hydrogen
             "r_inner": 1.e19,  # inner radius of cloud
             "r_in_pc": False,
-            "use_Q": True,
             "dust": False,
-            "efrac": -1.0,
-            "geometry": "sphere",
+            "efrac": 0.0,
             "abundance": "dopita",
+            "id_val": 1,
             "metals": [-1., -1., -1., -1., -1., -1., -1., -1., -1., -1., -1.]
             }
     for key, value in list(kwargs.items()):
@@ -89,40 +88,63 @@ def write_cloudy_input(**kwargs):
             f.write(to_print)
 
     # -----------------------------------------------------
+    _id = int(pars["id_val"])
+
     pars["gas_logZ"] = pars["logZ"]
     this_print('////////////////////////////////////')
     this_print('title {0}'.format(pars["model_name"].split('/')[-1]))
     this_print('////////////////////////////////////')
     this_print('set punch prefix "{0}"'.format(pars["model_name"]))
     this_print('print line precision 6')
+    
     cloudy_mod = "{}.mod".format(pars["model_name"])
     this_print('table star "{0}" {1}={2:.2e}'.format(cloudy_mod, "age", pars['age']))
-    if pars['use_Q']:
-        this_print('Q(H) = {0:.3f} log'.format(pars['logQ']))
-    else:
+    
+    # If DIG (id = 3) then use the ionization parameter else use the rate of ionizing photons
+    if _id == 3:
         this_print('ionization parameter = {0:.3f} log'.format(pars['logU']))
-
-    if pars['dust']:
-        this_print('grains ism {0:.2f} log no qheat'.format(pars['gas_logZ']))
-
+    else:
+        this_print('Q(H) = {0:.3f} log'.format(pars['logQ']))
+    
     if pars['r_in_pc']:
         pc_to_cm = 3.08568e18
         r_out = np.log10(pars['r_inner'] * pc_to_cm)
     else:
         r_out = np.log10(pars['r_inner'])
 
+    
     linefile = "cloudyLines.dat"
 
     # Check to see if there was an error in getting metallicities from the simulation
     # If so code revert back to using "dopita" abundances in place of "direct"
     if (any(q <= -1.0 for q in pars["metals"][1:]) and pars["abundance"] == "direct"):
         pars["abundance"] = "dopita"
-        print ("Warning: Unable to get metallicities from the simulation. Using \"dopita\" abundance instead")
+        print ("Warning: Unable to get metallicities from the simulation. This can be because you are binning the stars.\n \
+                Make sure that FORCE BINNED is set to False and set the max_age_direct appropriately to use this feature.\n \
+                Reverting to using abundances from \"dopita et al. 2001\"")
 
     if pars["abundance"] == "direct":
+        if pars['dust']:
+            this_print('grains orion {0:.2f} log no qheat'.format(pars['gas_logZ']))
+
         abund_el = ['He', 'C', 'N', 'O', 'Ne', 'Mg', 'Si', 'S', 'Ca', 'Fe']
         abund_metal = convert_metals(pars["metals"][1:])
         abund_str = "abundances "
+    
+        if cfg.par.FORCE_N_O_Pilyugin[_id]: # Adding N/O pilyugin relation
+            if 12 + abund_metal[3] < 8.14:
+                abund_metal[2] = -1.493 + abund_metal[3]
+            else:
+                abund_metal[2] =  1.489*(12 + abund_metal[3]) - 13.613 + abund_metal[3]
+       
+        if cfg.par.FORCE_N_O_ratio[_id]:
+            abund_metal[2] = cfg.par.N_O_ratio[_id] + abund_metal[3]
+
+        # Enhancing abundances for post-AGB stars
+        if _id == 1:
+            abund_metal[1] += cfg.par.PAGB_C_enhancement
+            abund_metal[2] += cfg.par.PAGB_N_enhancement
+
         for e in range(len(abund_el)):
             el_str = str(abund_el[e]) + " " + str(abund_metal[e]) + " "
             abund_str = abund_str + el_str
@@ -136,13 +158,44 @@ def write_cloudy_input(**kwargs):
                             pars["logZ"],
                             dust=pars["dust"],
                             re_z=False)
+
+        abund_C = float(abunds.elem_strs[1].split(" ")[3])
+        abund_N = float(abunds.elem_strs[2].split(" ")[3])
+        abund_O = float(abunds.elem_strs[3].split(" ")[3])
+        
+        if cfg.par.FORCE_N_O_Pilyugin[_id]: # Forcing the Nitrogen abundances to follow the N/O relation from Pilyugin et al. 2012
+            if 12 + abund_N < 8.14:
+                abund_N = -1.493 + abund_O
+            else:
+                abund_N =  1.489*(12 + abund_O) - 13.613 + abund_O
+
+        if cfg.par.FORCE_N_O_ratio[_id]:
+            abund_N = cfg.par.N_O_ratio[_id] + abund_O
+
+        # Enhancing abundances for post-AGB stars
+        if _id == 1:
+            abund_C += cfg.par.PAGB_C_enhancement
+            abund_N += cfg.par.PAGB_N_enhancement
+
+        abunds.elem_strs[1] = "element abundance carbon "+str(abund_C)+" log"
+        abunds.elem_strs[2] = "element abundance nitrogen "+str(abund_N)+" log"
+        abunds.elem_strs[3] = "element abundance oxygen "+str(abund_O)+" log"
+
         this_print(abunds.solarstr)
         for line in abunds.elem_strs:
             this_print(line)
     
-    this_print('radius {0:.3f} log'.format(r_out))
+    cf = 1 - pars['efrac']
+
+    # For DIG since we are making use of the ionization parameter we do need
+    # to set the inner radius, the geometry is plane parallel and not spherical
+    # and we are considering the covering factor to be 1. 
+    if _id != 3:
+        this_print('radius {0:.3f} log'.format(r_out))
+        this_print('sphere')
+        this_print('Covering factor {0:.3f}'.format(cf))
+    
     this_print('hden {0:.3f} log'.format(np.log10(pars['dens'])))
-    this_print('{}'.format(pars['geometry']))
     this_print('cosmic ray background')
     this_print('iterate to convergence max=5')
     this_print('stop temperature 100.0')
@@ -154,7 +207,7 @@ def write_cloudy_input(**kwargs):
     f.close()
 
 
-def get_output(model_name, dir_, qq, fsps_lam):
+def get_output(model_name, dir_, qq, fsps_lam, cell_width, id_val):
     lsun = 3.839e33
     c = constants.c.cgs.value * 1.e8
 
@@ -168,19 +221,32 @@ def get_output(model_name, dir_, qq, fsps_lam):
     wl = np.array([dat[0] for dat in wdat])
     sinds = np.argsort(wl)
     wl = wl[sinds]
+     
+    # For DIG we are making use of the ionization parameter (intensity case, see CLOUDY docs (Hazy 1))
+    # We need to multiply the output by the area of the cloud -- 6 sides of the cell in this case
+    if id_val == 3:
+        datflu = (datflu[sinds]*6*cell_width**2) / lsun 
 
-    datflu = datflu[sinds] / lsun / qq
+    else:
+        datflu = datflu[sinds] / lsun / qq
+    
     datflu = [float("{0:1.4e}".format(dat)) for dat in datflu]
+
     cont_data = np.genfromtxt(outcontfl, skip_header=1)
     nu = c / fsps_lam
     ang_0, diffuse_0 = cont_data[:, 0], cont_data[:, 2]
     ang_v, diffuse_in = air_to_vac(ang_0[::-1]), diffuse_0[::-1]
     diffuse_y = interp1d(ang_v, diffuse_in, fill_value=0.0, bounds_error=False)(fsps_lam)
-    diffuse_out = diffuse_y / nu / lsun / qq
+    
+    if id_val == 3:
+        diffuse_out = (diffuse_y*6*cell_width**2) / nu / lsun 
+    else:
+        diffuse_out = diffuse_y / nu / lsun / qq
+    
     return wl, datflu, fsps_lam, diffuse_out
 
 
-def clean_files(dir_, model_name, error=False):
+def clean_files(dir_, model_name, id_val, error=False):
     logging.info("Cleaning up temporary files")
     if not error:
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".out"))
@@ -188,16 +254,26 @@ def clean_files(dir_, model_name, error=False):
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".inicont"))
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".lin"))
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".outwcont"))
-        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".ascii"))
+        if id_val != 3:
+            os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".ascii"))
 
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".mod"))
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".in"))
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".out"))
+    if id_val != 3:
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".mod"))
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".in"))
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".out"))
 
 
-def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals, Dust=False, abund="dopita", useq=True, clean_up=True):
+def get_nebular(spec_lambda, sspi, nh, Metals, logq = 0.0, radius = 1.e19, logu = 0.0, logz = 0.0, logq_1=0.0, 
+        Cell_width=1, Dust=False, abund="dopita", clean_up=True, index=1, efrac=0.0):
+    
+    # Since we don't need the logQs for DIG as we are using 
+    # the ionization parmeter instead. Set them to 0 just as a place holder 
+    if index == 3:
+        logq = 0.0
+        logq_1 = 0.0
+    
     nspec = len(spec_lambda)
-    frac_obrun = 0.0
+    frac_obrun = efrac
     clight = constants.c.cgs.value*1.e8
     nebular_smooth_init = 0
 
@@ -206,20 +282,20 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
     model_name = filename.split(".")[0]
 
     logging.info("Writing CLOUDY input file")
-    dir_= cfg.par.pd_source_dir + "/powderday/nebular_emission"
+    dir_= cfg.par.pd_source_dir + "powderday/nebular_emission"
     dir_base = os.getcwd()
-
     write_cloudy_input(dir_=dir_,
-                       model_name=model_name,
-                       r_inner=radius,
-                       dens=nh,
-                       use_Q=useq,
-                       logQ=logq,
-                       logU=logu,
-                       logZ=logz,
-                       abundance=abund,
-                       dust=Dust,
-                       metals=Metals)
+                       model_name = model_name,
+                       r_inner = radius,
+                       dens = nh,
+                       logQ = logq,
+                       logU = logu,
+                       logZ = logz,
+                       abundance = abund,
+                       dust = Dust,
+                       id_val = index,
+                       efrac = frac_obrun,
+                       metals = Metals)
 
     logging.info("Input SED file written")
     logging.info("Running CLOUDY")
@@ -229,18 +305,20 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
     f_out = open(dir_ + "/temp_files/"+model_name+".out", 'r')
     content = f_out.readlines()
     check = np.all(['OK' in content[-1]])
+    
     if check:
         logging.info("CLOUDY run finished sucessfully")
 
     else:
-        logging.info("CLOUDY run was unsucessful")
+        print ("WARNING: The CLOUDY run was unsucessful. Please see the cloudy output file "+ model_name+".out"+
+               " located in "+ dir_ + "/temp_files/"+" to figure out why this happened.")
         if clean_up:
-            clean_files(dir_, model_name, error=True)
+            clean_files(dir_, model_name, index, error=True)
         raise ValueError('CLOUDY run was unsucessful')
 
     logging.info("Getting output spectrum")
 
-    nebem_line_pos, nebem_line, readlambneb, readcontneb = get_output(model_name, dir_, 10**logq, spec_lambda)
+    nebem_line_pos, nebem_line, readlambneb, readcontneb = get_output(model_name, dir_, 10**logq, spec_lambda, Cell_width, index)
     nebem_cont = interp1d(readlambneb, np.log10(readcontneb + 10 ** (-95.0)),
                           fill_value=-95.0, bounds_error=False)(spec_lambda)
 
@@ -267,8 +345,8 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
     sspo = sspo + 10 ** nebem_cont * (10**logq_1)
     for i in range(nemline):
         sspo = sspo + nebem_line[i] * gaussnebarr[i] * (10**logq_1)
-
+        
     if clean_up:
-        clean_files(dir_, model_name,  error=False)
+        clean_files(dir_, model_name, index, error=False)
 
-    return sspo
+    return sspo, nebem_line_pos, np.array(nebem_line)*(10**logq_1)
